refactor: move definitions.py inside the module code
