+++ museek/plugin/demo/demo_load_plugin.py
@@ -4,7 +4,7 @@ from io import BytesIO
 import requests
 from PIL import Image
 
-from definitions import ROOT_DIR
+from museek.definitions import ROOT_DIR
 from ivory.plugin.abstract_plugin import AbstractPlugin
 from ivory.utils.result import Result
 from museek.enums.demo_enum import DemoEnum
