+++ museek/plugin/in_plugin.py
@@ -1,7 +1,7 @@
 import os
 from datetime import datetime
 
-from definitions import ROOT_DIR
+from museek.definitions import ROOT_DIR
 from ivory.plugin.abstract_plugin import AbstractPlugin
 from ivory.utils.result import Result
 from museek.enums.result_enum import ResultEnum
